requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==0.28.1
gunicorn==21.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0 